refactor(glossary): move 81-term parity server-side to fix bundle budget
The first cut added all 81 missing terms to lib/glossary.ts, but that file is
client-bundled (JargonTooltip imports GLOSSARY), so it pushed the shared client
chunk +49.5 kB over the 12 MB bundle budget (CI caught it).

Keep lib/glossary.ts lean at 122 (the client tooltip set — now identical to
main, zero bundle change) and move the 81 specialised terms to a new
server-only lib/glossary-extended.ts exporting FULL_GLOSSARY_ENTRIES (203).
Point the server-side consumers at the full set:

- app/sitemap.ts — now emits all 203 /glossary/<slug> URLs (was missing the 81
  live term pages, a real discoverability gap for AI/search).
- lib/keyword-linking.ts — 203 internal-link targets in article output.
- lib/glossary-db.ts — complete build-time fallback when the DB is unreachable.

Net: the GEO win (full sitemap coverage + more internal links) with no client
cost. Verified no Client Component imports glossary-extended, so it stays out
of the client graph. 18 glossary tests green; tsc clean on source.
