+++ docs/strategy/FIN_NOTEBOOK.md
@@ -71,11 +71,17 @@ is a lens applied to existing work, not a new build — most of the moats alread
 - [x] **Glossary "→200" was ALREADY met — corrected 2026-05-21.** The live glossary is
   DB-backed (`public.glossary_terms`, seeded by migration `20260419` with **203 terms**);
   `lib/glossary.ts` (122) is only the `JargonTooltip` source + build-time fallback and had
-  drifted behind. Real gap was static/DB parity, not new content — synced the 81 missing
-  terms into `lib/glossary.ts` (now 203, a clean superset match; 0 static-only terms).
-  No new content authored / no prod migration needed. Lesson (again): grep the runtime
-  source before scoping from a tracker count. *Net-new terms beyond 203 are a separate,
-  genuinely-new-content task if ever wanted — route through the content loop for accuracy.*
+  drifted behind. Real gap was static/DB parity, not new content. First attempt synced all
+  81 into `lib/glossary.ts` — but that file is **client-bundled** (JargonTooltip imports
+  `GLOSSARY`), so it tipped the shared client chunk +49.5 kB over the 12 MB bundle budget
+  (CI caught it). **Final shape (#1156):** keep `lib/glossary.ts` lean at 122 (client tooltip
+  set, == main, zero bundle change); put the 81 specialised terms in a **server-only**
+  `lib/glossary-extended.ts` (`FULL_GLOSSARY_ENTRIES` = 203) that powers the sitemap (was
+  missing 81 live term pages → now complete), internal-link targets, and the DB fallback.
+  Net: a GEO win (all 203 term pages in the sitemap + more internal-link targets) with no
+  client cost. Lesson (again): grep the runtime source before scoping from a tracker count,
+  and watch what's client- vs server-bundled. *Net-new terms beyond 203 = separate
+  new-content task; route through the content loop for accuracy.*
 - [ ] Confirm QQ public Q&A pages emit FAQ + (where apt) Speakable schema.
 
 **Deeper GEO dive (2026-05-21) — what else to add/adjust, grounded in the codebase:**
